Artifacts can no longer be destroyed by the 'k' option of the [y/n/k] prompt.

+++ zangband/src/cmd1.c
@@ -875,8 +875,20 @@ void carry(int pickup)
 					
 					if ((i == 'K') || (i == 'k'))
 					{
-						/* Destroy the object */
-						delete_object_idx(this_o_idx);
+						/* Artifact? */
+						if (!can_player_destroy_object(o_ptr))
+						{
+							/* Describe the object (with {terrible/special}) */
+							object_desc(o_name, o_ptr, TRUE, 3);
+
+							/* Message */
+							msg_format("You cannot destroy the %s.", o_name);
+						}
+						else
+						{
+							/* Destroy the object */
+							delete_object_idx(this_o_idx);
+						}
 					}
 				}
